refactor(shared): promote CalendarDay type to app/shared (#983)
