+++ app/components/financial-calendar/FinancialCalendar/FinancialCalendar.vue
@@ -2,10 +2,8 @@
 import { TrendingDown, TrendingUp, ChevronLeft, ChevronRight } from "lucide-vue-next";
 import { computed, watch } from "vue";
 import type { FinancialCalendarProps, FinancialCalendarEmits } from "./FinancialCalendar.types";
-import {
-  useFinancialCalendar,
-  type CalendarDay,
-} from "~/features/transactions/composables/useFinancialCalendar";
+import { useFinancialCalendar } from "~/features/transactions/composables/useFinancialCalendar";
+import type { CalendarDay } from "~/shared/types/financial-calendar";
 import { formatCurrency } from "~/utils/currency";
 
 const props = withDefaults(defineProps<FinancialCalendarProps>(), {
